requirements.txt
fastapi
uvicorn
twilio
openai
python-decouple
sqlalchemy
psycopg2-binary
python-multipart
pyngrok
# gradio_client
gradio-client @ git+https://github.com/gradio-app/gradio@e5292fa86d9a1f641bff409b12046161e7159418#subdirectory=client/python
https://gradio-pypi-previews.s3.amazonaws.com/e5292fa86d9a1f641bff409b12046161e7159418/gradio-5.1.0-py3-none-any.whl
numpy==1.21.6
# Pillow==9.4.0
azure-storage-blob==12.13.1
azure-cognitiveservices-vision-computervision==0.9.0
opencv-python-headless==4.7.0.68
opencv-python==4.7.0.68
